Register page now does create user registry in database and fields have validations.

+++ server/src/main/java/com/knowy/server/service/TokenService.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 @Service
 public class TokenService {
 
-	public String createPasswordResetToken(String email, Long userId) {
+	public String createPasswordResetToken(String email, Integer userId) {
 		// TODO - Implementar JWT
 		return UUID.randomUUID().toString();
 	}
